asterisk: Band aid for a crash on RTP RTCP message
This might be a general TLS issue with Asterisk on ARM. I have not
looked into why this is happening. Create a band-aid.

Crash:
(gdb) bt
it crashed when formating/printing the &addr in the rtcp recv method
...

RTCP:
Packet generated from Ekiga 3.2.7 (Debian Unstable).

Cause:
Not very clear, maybe something with TLS handling of asterisk