misc: Modernize and call it packagegroup-
The notion of task- has been deprecated a long time ago. Call
it packagegroup now and for "packagegroup-osmocom" provide the
compability. This has not been compile tested.

The edison build will require a bbappend to provide packagegroup
now but I will do this once I need to build for Edison again.