osmo-ggsn: Drop untested sysvinit support
These bits come from times were sysvinit was used in stead of systemd in
yocto. Nowadays nobody is known to be using the recipes with sysvinit
instead of systemd, and anyway most osmocom recipes doesn't come with
sysvinit support, so let's drop it. If sysvinit support is required, it
should be added through autotools install process instead.

Change-Id: I303ea210e14e0e6e9a18b1bb4c25d470924d80c6