openbsc: Drop untested sysvinit support
These bits come from times were sysvinit was used in stead of systemd in
yocto. Nowadays nobody is known to be using the recipes with sysvinit
instead of systemd, and anyway most osmocom recipes doesn't come with
sysvinit support, so let's drop it. If sysvinit support is required, it
should be added through autotools install process instead.

As update-rc.d was removed, it was noticied that systemd class inherit
was missing, so this commit adds it. As a result, it's not required to
define systemd service file paths in FILES_*.

Change-Id: I864e4a61aed7933c46b80aa26598234faba43395